requirements.txt
cffi==2.0.0
cryptography==46.0.5
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
pytest==9.0.2
fastapi==0.111.1
uvicorn[standard]==0.30.3
aiokafka==0.10.0
python-dotenv==1.0.1